Revert "seperate method and class"
This reverts commit 467cc9ab0b612db42acc18984253507efda3a03d.

# Conflicts:
#	src/Analyzers/MSTest.Analyzers/PublicAPI.Unshipped.txt
