+++ src/Analyzers/MSTest.Analyzers/Helpers/DiagnosticIds.cs
@@ -39,6 +39,5 @@ internal static class DiagnosticIds
     public const string ReviewAlwaysTrueAssertConditionAnalyzerRuleId = "MSTEST0032";
     public const string NonNullableReferenceNotInitializedSuppressorRuleId = "MSTEST0033";
     public const string UseClassCleanupBehaviorEndOfClassRuleId = "MSTEST0034";
-    public const string UseDeploymentItemWithTestClassRuleId = "MSTEST0035";
-    public const string UseDeploymentItemWithTestMethodRuleId = "MSTEST0036";
+    public const string UseDeploymentItemWithTestMethodOrTestClassRuleId = "MSTEST0035";
 }
