+++ src/Analyzers/MSTest.Analyzers/UseDeploymentItemWithTestMethodOrTestClassAnalyzer.cs
@@ -13,17 +13,17 @@ using MSTest.Analyzers.Helpers;
 namespace MSTest.Analyzers;
 
 /// <summary>
-/// MSTEST0036: <inheritdoc cref="Resources.UseDeploymentItemWithTestMethodOrTestClassTitle"/>.
+/// MSTEST0034: <inheritdoc cref="Resources.UseDeploymentItemWithTestMethodOrTestClassTitle"/>.
 /// </summary>
 [DiagnosticAnalyzer(LanguageNames.CSharp, LanguageNames.VisualBasic)]
-public sealed class UseDeploymentItemWithTestMethodAnalyzer : DiagnosticAnalyzer
+public sealed class UseDeploymentItemWithTestMethodOrTestClassAnalyzer : DiagnosticAnalyzer
 {
     private static readonly LocalizableResourceString Title = new(nameof(Resources.UseDeploymentItemWithTestMethodOrTestClassTitle), Resources.ResourceManager, typeof(Resources));
     private static readonly LocalizableResourceString Description = new(nameof(Resources.UseDeploymentItemWithTestMethodOrTestClassDescription), Resources.ResourceManager, typeof(Resources));
     private static readonly LocalizableResourceString MessageFormat = new(nameof(Resources.UseDeploymentItemWithTestMethodOrTestClassMessageFormat), Resources.ResourceManager, typeof(Resources));
 
     internal static readonly DiagnosticDescriptor UseDeploymentItemWithTestMethodOrTestClassRule = DiagnosticDescriptorHelper.Create(
-        DiagnosticIds.UseDeploymentItemWithTestMethodRuleId,
+        id: DiagnosticIds.UseDeploymentItemWithTestMethodOrTestClassRuleId,
         Title,
         MessageFormat,
         Description,
@@ -42,36 +42,54 @@ public sealed class UseDeploymentItemWithTestMethodAnalyzer : DiagnosticAnalyzer
         context.RegisterCompilationStartAction(context =>
         {
             if (context.Compilation.TryGetOrCreateTypeByMetadataName(WellKnownTypeNames.MicrosoftVisualStudioTestToolsUnitTestingTestMethodAttribute, out INamedTypeSymbol? testMethodAttributeSymbol)
+                 && context.Compilation.TryGetOrCreateTypeByMetadataName(WellKnownTypeNames.MicrosoftVisualStudioTestToolsUnitTestingTestClassAttribute, out INamedTypeSymbol? testClassAttributeSymbol)
                  && context.Compilation.TryGetOrCreateTypeByMetadataName(WellKnownTypeNames.MicrosoftVisualStudioTestToolsUnitTestingDeploymentItemAttribute, out INamedTypeSymbol? deploymentItemAttributeSymbol))
             {
                 context.RegisterSymbolAction(
-                    context => AnalyzeSymbol(context, testMethodAttributeSymbol, deploymentItemAttributeSymbol),
-                    SymbolKind.Method);
+                    context => AnalyzeSymbol(context, testMethodAttributeSymbol, testClassAttributeSymbol, deploymentItemAttributeSymbol),
+                    new[] { SymbolKind.NamedType, SymbolKind.Method });
             }
         });
     }
 
-    private static void AnalyzeSymbol(SymbolAnalysisContext context, INamedTypeSymbol testMethodAttributeSymbol, INamedTypeSymbol deploymentItemAttributeSymbol)
+    private static void AnalyzeSymbol(SymbolAnalysisContext context, INamedTypeSymbol testMethodAttributeSymbol, INamedTypeSymbol testClassAttributeSymbol, INamedTypeSymbol deploymentItemAttributeSymbol)
     {
-        var methodSymbol = (IMethodSymbol)context.Symbol;
+        var namedTypeSymbol = (INamedTypeSymbol)context.Symbol;
 
         bool hasDeploymentItemAttribute = false;
         bool isTestMethodOrTestClass = false;
-        foreach (AttributeData attribute in methodSymbol.GetAttributes())
+        if (namedTypeSymbol.TypeKind == TypeKind.Class)
         {
-            if (attribute.AttributeClass.Inherits(testMethodAttributeSymbol))
+            foreach (AttributeData attribute in namedTypeSymbol.GetAttributes())
             {
-                isTestMethodOrTestClass = true;
+                if (attribute.AttributeClass.Inherits(testClassAttributeSymbol))
+                {
+                    isTestMethodOrTestClass = true;
+                }
+                else if (SymbolEqualityComparer.Default.Equals(attribute.AttributeClass, deploymentItemAttributeSymbol))
+                {
+                    hasDeploymentItemAttribute = true;
+                }
             }
-            else if (SymbolEqualityComparer.Default.Equals(attribute.AttributeClass, deploymentItemAttributeSymbol))
+        }
+        else if (context.Symbol is IMethodSymbol methodSymbol)
+        {
+            foreach (AttributeData attribute in methodSymbol.GetAttributes())
             {
-                hasDeploymentItemAttribute = true;
+                if (attribute.AttributeClass.Inherits(testMethodAttributeSymbol))
+                {
+                    isTestMethodOrTestClass = true;
+                }
+                else if (SymbolEqualityComparer.Default.Equals(attribute.AttributeClass, deploymentItemAttributeSymbol))
+                {
+                    hasDeploymentItemAttribute = true;
+                }
             }
         }
 
         if (hasDeploymentItemAttribute && !isTestMethodOrTestClass)
         {
-            context.ReportDiagnostic(methodSymbol.CreateDiagnostic(UseDeploymentItemWithTestMethodOrTestClassRule));
+            context.ReportDiagnostic(namedTypeSymbol.CreateDiagnostic(UseDeploymentItemWithTestMethodOrTestClassRule));
         }
     }
 }
