--- test/UnitTests/MSTest.Analyzers.UnitTests/UseDeploymentItemWithTestClassAnalyzerTests.cs
@@ -1,27 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using VerifyCS = MSTest.Analyzers.Test.CSharpCodeFixVerifier<
-    MSTest.Analyzers.UseDeploymentItemWithTestClassAnalyzer,
-    Microsoft.CodeAnalysis.Testing.EmptyCodeFixProvider>;
-
-namespace MSTest.Analyzers.Test;
-
-[TestGroup]
-public sealed class UseDeploymentItemWithTestClassAnalyzerTests(ITestExecutionContext testExecutionContext) : TestBase(testExecutionContext)
-{
-    public async Task WhenTestClassHasDeploymentItem_NoDiagnostic()
-    {
-        string code = """
-            using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-            [TestClass]
-            [DeploymentItem("")]
-            public class MyTestClass
-            {
-            }
-            """;
-
-        await VerifyCS.VerifyAnalyzerAsync(code);
-    }
-}
