+++ test/UnitTests/MSTest.Analyzers.UnitTests/UseDeploymentItemWithTestMethodOrTestClassAnalyzerTests.cs
@@ -2,14 +2,29 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using VerifyCS = MSTest.Analyzers.Test.CSharpCodeFixVerifier<
-    MSTest.Analyzers.UseDeploymentItemWithTestMethodAnalyzer,
+    MSTest.Analyzers.UseDeploymentItemWithTestMethodOrTestClassAnalyzer,
     Microsoft.CodeAnalysis.Testing.EmptyCodeFixProvider>;
 
 namespace MSTest.Analyzers.Test;
 
 [TestGroup]
-public sealed class UseDeploymentItemWithTestMethodAnalyzerTests(ITestExecutionContext testExecutionContext) : TestBase(testExecutionContext)
+public sealed class UseDeploymentItemWithTestMethodOrTestClassAnalyzerTests(ITestExecutionContext testExecutionContext) : TestBase(testExecutionContext)
 {
+    public async Task WhenTestClassHasDeploymentItem_NoDiagnostic()
+    {
+        string code = """
+            using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+            [TestClass]
+            [DeploymentItem("")]
+            public class MyTestClass
+            {
+            }
+            """;
+
+        await VerifyCS.VerifyAnalyzerAsync(code);
+    }
+
     public async Task WhenTestMethodHasDeploymentItem_NoDiagnostic()
     {
         string code = """
@@ -18,6 +33,7 @@ public sealed class UseDeploymentItemWithTestMethodAnalyzerTests(ITestExecutionC
             [TestClass]
             public class MyTestClass
             {
+                [TestMethod]
                 [DeploymentItem("")]
                 public void MyTestMethod()
                 {
